+++ R/bayesopt_ego.R
@@ -129,7 +129,7 @@ bayesopt_ego = function(
     init_design_size = 4L * search_space$length
   }
   if (!is.null(init_design_size) && instance$archive$n_evals == 0L) {
-    design = generate_design_sobol(search_space, n = init_design_size)$data
+    design = generate_design_random(search_space, n = init_design_size)$data
     instance$eval_batch(design)
   }
 
@@ -145,6 +145,7 @@ bayesopt_ego = function(
       if (isTRUE((instance$archive$n_evals - init_design_size + 1L) %% random_interleave_iter == 0)) {
         stop(set_class(list(message = "Random interleaving", call = NULL), classes = c("random_interleave", "mbo_error", "error", "condition")))
       }
+
       acq_function$surrogate$update()
       acq_function$update()
       acq_optimizer$optimize()