+++ tests/testthat/test_OptimizerAsyncMbo.R
@@ -15,7 +15,7 @@ test_that("OptimizerAsyncMbo works in defaults", {
 
   expect_data_table(optimizer$optimize(instance), nrows = 1L)
   expect_data_table(instance$archive$data, min.rows = 10L)
-  expect_names(names(instance$archive$data), must.include = c("acq_cb", ".already_evaluated", "acq_lambda_0", "acq_lambda"))
+  expect_names(names(instance$archive$data), must.include = c("acq_cb")) # , ".already_evaluated", "acq_lambda_0", "acq_lambda"
 
   expect_rush_reset(instance$rush)
 })