+++ tests/testthat/test_SurrogateLearner.R
@@ -11,8 +11,10 @@ test_that("SurrogateLearner API works", {
 
   xdt = data.table(x = seq(-1, 1, length.out = 5L))
   pred = surrogate$predict(xdt)
-  expect_data_table(pred, col.names = "named", nrows = 5, ncols = 2, any.missing = FALSE)
-  expect_named(pred, c("mean", "se"))
+  expect_list(pred, len = 2L)
+  expect_names(names(pred), identical.to = c("mean", "se"))
+  expect_numeric(pred$mean, len = 5L)
+  expect_numeric(pred$se, len = 5L)
 
   # upgrading error class works
   surrogate = SurrogateLearner$new(LearnerRegrError$new(), archive = inst$archive)